requirements.txt
telethon==1.41.2
cryptg==0.5.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
base58==2.1.1

